fix: fix widget background color
