+++ SNUTT-2022/SNUTT/Assets/Assets.xcassets/Colors/tab.background.colorset/Contents.json
@@ -22,7 +22,7 @@
       "color" : {
         "color-space" : "display-p3",
         "components" : {
-          "alpha" : "0.500",
+          "alpha" : "0.300",
           "blue" : "0.000",
           "green" : "0.000",
           "red" : "0.000"
