+++ SNUTT-2022/SNUTTWidget/Assets.xcassets/WidgetBackground.colorset/Contents.json
@@ -1,6 +1,28 @@
 {
   "colors" : [
     {
+      "color" : {
+        "platform" : "ios",
+        "reference" : "systemBackgroundColor"
+      },
+      "idiom" : "universal"
+    },
+    {
+      "appearances" : [
+        {
+          "appearance" : "luminosity",
+          "value" : "dark"
+        }
+      ],
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "alpha" : "1.000",
+          "blue" : "0.169",
+          "green" : "0.169",
+          "red" : "0.169"
+        }
+      },
       "idiom" : "universal"
     }
   ],
